Alternate printing option
Conky execgraph requires an integer value between 0 and 100.  A "-p"
command line switch (-n was already in use) can now be used to provide
an integer representation of the amplitude, so this can be called from
Conky.

Recommend calling by creating a shell script like so:

pastat -p -n 1 -q -s "SINK NAME"

Get conky to use this script in execigraph and you can have a (slow)
visualizer for your audio in conky!